requirements.txt
Django==1.11.4
djangorestframework==3.6.3
MySQL-python==1.2.5
olefile==0.44
Pillow==4.2.1
pytz==2017.2
wheel==0.24.0
